use logMonoTime of the message

+++ can/common.pxd
@@ -60,7 +60,7 @@ cdef extern from "common_dbc.h":
 
   cdef struct SignalValue:
     uint32_t address
-    uint16_t ts
+    uint64_t ts
     const char* name
     double value
