+++ can/common_dbc.h
@@ -24,7 +24,7 @@ struct MessageParseOptions {
 
 struct SignalValue {
   uint32_t address;
-  uint16_t ts;
+  uint64_t ts;
   const char* name;
   double value;
 };
